+++ accounting/src/apideck_accounting_unify/models/accounting_taxratesupdateop.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 from .httpmetadata import HTTPMetadata, HTTPMetadataTypedDict
-from .taxrate import TaxRateInput, TaxRateInputTypedDict
+from .taxrate_input import TaxRateInput, TaxRateInputTypedDict
 from .unexpectederrorresponse import (
     UnexpectedErrorResponse,
     UnexpectedErrorResponseTypedDict,