+++ accounting/src/apideck_accounting_unify/models/companyinfo.py
@@ -6,7 +6,7 @@ from .currency import Currency
 from .custommappings import CustomMappings, CustomMappingsTypedDict
 from .email import Email, EmailTypedDict
 from .phonenumber import PhoneNumber, PhoneNumberTypedDict
-from .taxrate import TaxRate, TaxRateTypedDict
+from .taxrate_input import TaxRate, TaxRateTypedDict
 from apideck_accounting_unify.types import (
     BaseModel,
     Nullable,